requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Groq API
groq==0.4.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.1

# Vector/Embeddings
voyageai==0.2.1  # or openai for embeddings
sentence-transformers==2.3.1  # Alternative local embeddings

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.2.0

# RAG Framework (choose one)
langchain==0.1.4


# Text Processing
tiktoken==0.6.0
beautifulsoup4==4.12.3

# Celery & Redis
celery==5.3.6
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
black==24.1.1
ruff==0.2.0